srcs/back/requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
tzdata==2022.7
whitenoise==6.7.0
channels==3.0.5
channels_redis
psycopg2-binary
djangorestframework
django-cors-headers
PyJWT
pytz
python-dotenv
requests
pyotp
qrcode
pillow